Extract new app resource form to a new independant component
